+++ HISTORY.md
@@ -1,6 +1,13 @@
 Release History
 ===============
 
+0.5.0 (2023-11-09)
+------------------
+
+**Improvements**
+- Add --wait and --timeout for queries run.
+- Add option to supply query input as text.
+
 0.3.0 (2023-11-07)
 ------------------